move layout out of routes dir
